requirements.txt
git+https://github.com/openkim/kim-tools.git
chgnet
matgl
mace-torch
